rename: LoadWorkerHandle.ts -> VolumeLoaderContext.ts
