+++ public/index.ts
@@ -21,7 +21,7 @@ import {
 import { OpenCellLoader } from "../src/loaders/OpenCellLoader";
 import { State, TestDataSpec } from "./types";
 import { getDefaultImageInfo } from "../src/Volume";
-import VolumeLoaderContext from "../src/workers/LoadWorkerHandle";
+import VolumeLoaderContext from "../src/workers/VolumeLoaderContext";
 import { DATARANGE_UINT8 } from "../src/types";
 
 const CACHE_MAX_SIZE = 1_000_000_000;
